release: v0.2.0 — WASM package, vendored parity guard, manifest sync
- Add @jsonstat-validator/wasm package (wasm-pack wrapper) with parity test

- Add crates/validator vendored_parity test guarding the _vendored snapshot

- Sync rules-manifest.json and the vendored manifest/schemas snapshot

- Bump TS / CLI / Rust / WASM packages to 0.2.0

- Update CI workflow, READMEs, and .gitignore

+++ crates/validator/src/types.rs
@@ -60,8 +60,10 @@ pub struct ResultMeta {
     pub rule_set_version: String,
     #[serde(rename = "schemaVersion")]
     pub schema_version: String,
+    /// Validation wall-clock time in ms (0 on the Wasm surface, which has no monotonic clock).
+    /// u64 (not u128) keeps the value serde-wasm-bindgen-safe.
     #[serde(rename = "durationMs")]
-    pub duration_ms: u128,
+    pub duration_ms: u64,
 }
 
 #[derive(Debug, Clone, Serialize)]
